refactor(coding-agent): extract gh search module
Move the search-op domain (issue/pr/code/commit/repo search, date-bound parsing, API-to-result mappers) into gh-search; facade re-exports the public search helpers.

Op: compress
